remove unused files

+++ src/main/resources/db/migration/V1__create_inventory_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE inventory (
-    id         NUMBER(19)  GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    product_id NUMBER(19)  NOT NULL,
-    quantity   NUMBER(10)  DEFAULT 0 NOT NULL,
+    id         BIGINT  GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    product_id BIGINT  NOT NULL,
+    quantity   INTEGER DEFAULT 0 NOT NULL,
     CONSTRAINT uq_inventory_product_id UNIQUE (product_id)
 );
